TransportSource Issues
Fixed an issue where when Reference was not on, the pause and play button felt laggy. Also now when you click play it will not stop playing when going to Reference, it will only stop if you click pause or stop. Now you can do both smoothly even though it's not on Reference